requirements.txt
python-crontab==3.0.0
customtkinter==5.2.2
rumps==0.4.0
